+++ flight/PiOS/Common/pios_mpu6050.c
@@ -476,32 +476,32 @@ static void PIOS_MPU6050_Task(void* parameters)
 			break;
 
 		case PIOS_MPU60X0_TOP_90DEG:
-			accel_data.y = (int16_t) - (mpu6050_rec_buf[IDX_ACCEL_YOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_YOUT_L]);
+			accel_data.y = - (int16_t)(mpu6050_rec_buf[IDX_ACCEL_YOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_YOUT_L]);
 			accel_data.x = (int16_t)(mpu6050_rec_buf[IDX_ACCEL_XOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_XOUT_L]);
-			gyro_data.y  = (int16_t) - (mpu6050_rec_buf[IDX_GYRO_YOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_YOUT_L]);
+			gyro_data.y  = - (int16_t)(mpu6050_rec_buf[IDX_GYRO_YOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_YOUT_L]);
 			gyro_data.x  = (int16_t)(mpu6050_rec_buf[IDX_GYRO_XOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_XOUT_L]);
 			break;
 
 		case PIOS_MPU60X0_TOP_180DEG:
-			accel_data.y = (int16_t) - (mpu6050_rec_buf[IDX_ACCEL_XOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_XOUT_L]);
-			accel_data.x = (int16_t) - (mpu6050_rec_buf[IDX_ACCEL_YOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_YOUT_L]);
-			gyro_data.y  = (int16_t) - (mpu6050_rec_buf[IDX_GYRO_XOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_XOUT_L]);
-			gyro_data.x  = (int16_t) - (mpu6050_rec_buf[IDX_GYRO_YOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_YOUT_L]);
+			accel_data.y = - (int16_t)(mpu6050_rec_buf[IDX_ACCEL_XOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_XOUT_L]);
+			accel_data.x = - (int16_t)(mpu6050_rec_buf[IDX_ACCEL_YOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_YOUT_L]);
+			gyro_data.y  = - (int16_t)(mpu6050_rec_buf[IDX_GYRO_XOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_XOUT_L]);
+			gyro_data.x  = - (int16_t)(mpu6050_rec_buf[IDX_GYRO_YOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_YOUT_L]);
 			break;
 
 		case PIOS_MPU60X0_TOP_270DEG:
 			accel_data.y = (int16_t)(mpu6050_rec_buf[IDX_ACCEL_YOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_YOUT_L]);
-			accel_data.x = (int16_t) - (mpu6050_rec_buf[IDX_ACCEL_XOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_XOUT_L]);
+			accel_data.x = - (int16_t)(mpu6050_rec_buf[IDX_ACCEL_XOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_XOUT_L]);
 			gyro_data.y  = (int16_t)(mpu6050_rec_buf[IDX_GYRO_YOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_YOUT_L]);
-			gyro_data.x  = (int16_t) - (mpu6050_rec_buf[IDX_GYRO_XOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_XOUT_L]);
+			gyro_data.x  = - (int16_t)(mpu6050_rec_buf[IDX_GYRO_XOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_XOUT_L]);
 			break;
 		}
 
-		gyro_data.z  = (int16_t) - (mpu6050_rec_buf[IDX_GYRO_ZOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_ZOUT_L]);
-		accel_data.z = (int16_t) - (mpu6050_rec_buf[IDX_ACCEL_ZOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_ZOUT_L]);
+		gyro_data.z  = - (int16_t)(mpu6050_rec_buf[IDX_GYRO_ZOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_ZOUT_L]);
+		accel_data.z = - (int16_t)(mpu6050_rec_buf[IDX_ACCEL_ZOUT_H] << 8 | mpu6050_rec_buf[IDX_ACCEL_ZOUT_L]);
 
-		int16_t raw_temp = (mpu6050_rec_buf[IDX_TEMP_OUT_H] << 8 | mpu6050_rec_buf[IDX_TEMP_OUT_L]);
-		float temperature = 35.0f + ((float) raw_temp + 512.0f) / 340.0f;
+		int16_t raw_temp = (int16_t)(mpu6050_rec_buf[IDX_TEMP_OUT_H] << 8 | mpu6050_rec_buf[IDX_TEMP_OUT_L]);
+		float temperature = 35.0f + ((float)raw_temp + 512.0f) / 340.0f;
 
 		// Apply sensor scaling
 		float accel_scale = PIOS_MPU6050_GetAccelScale();
@@ -517,10 +517,10 @@ static void PIOS_MPU6050_Task(void* parameters)
 		gyro_data.temperature = temperature;
 
 		portBASE_TYPE xHigherPriorityTaskWoken_accel;
-		xQueueSendToBackFromISR(dev->accel_queue, (void*) &accel_data, &xHigherPriorityTaskWoken_accel);
+		xQueueSendToBackFromISR(dev->accel_queue, (void*)&accel_data, &xHigherPriorityTaskWoken_accel);
 
 		portBASE_TYPE xHigherPriorityTaskWoken_gyro;
-		xQueueSendToBackFromISR(dev->gyro_queue, (void*) &gyro_data, &xHigherPriorityTaskWoken_gyro);
+		xQueueSendToBackFromISR(dev->gyro_queue, (void*)&gyro_data, &xHigherPriorityTaskWoken_gyro);
 
 #else
 
@@ -534,25 +534,25 @@ static void PIOS_MPU6050_Task(void* parameters)
 			break;
 
 		case PIOS_MPU60X0_TOP_90DEG:
-			gyro_data.y  = (int16_t) - (mpu6050_rec_buf[IDX_GYRO_YOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_YOUT_L]);
+			gyro_data.y  = - (int16_t)(mpu6050_rec_buf[IDX_GYRO_YOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_YOUT_L]);
 			gyro_data.x  = (int16_t)(mpu6050_rec_buf[IDX_GYRO_XOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_XOUT_L]);
 			break;
 
 		case PIOS_MPU60X0_TOP_180DEG:
-			gyro_data.y  = (int16_t) - (mpu6050_rec_buf[IDX_GYRO_XOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_XOUT_L]);
-			gyro_data.x  = (int16_t) - (mpu6050_rec_buf[IDX_GYRO_YOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_YOUT_L]);
+			gyro_data.y  = - (int16_t)(mpu6050_rec_buf[IDX_GYRO_XOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_XOUT_L]);
+			gyro_data.x  = - (int16_t)(mpu6050_rec_buf[IDX_GYRO_YOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_YOUT_L]);
 			break;
 
 		case PIOS_MPU60X0_TOP_270DEG:
 			gyro_data.y  = (int16_t)(mpu6050_rec_buf[IDX_GYRO_YOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_YOUT_L]);
-			gyro_data.x  = (int16_t) - (mpu6050_rec_buf[IDX_GYRO_XOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_XOUT_L]);
+			gyro_data.x  = - (int16_t)(mpu6050_rec_buf[IDX_GYRO_XOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_XOUT_L]);
 			break;
 		}
 
-		gyro_data.z = (int16_t) - (mpu6050_rec_buf[IDX_GYRO_ZOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_ZOUT_L]);
+		gyro_data.z = - (int16_t)(mpu6050_rec_buf[IDX_GYRO_ZOUT_H] << 8 | mpu6050_rec_buf[IDX_GYRO_ZOUT_L]);
 
-		int32_t raw_temp = (mpu6050_rec_buf[IDX_TEMP_OUT_H] << 8 | mpu6050_rec_buf[IDX_TEMP_OUT_L]);
-		float temperature = 35.0f + ((float) raw_temp + 512.0f) / 340.0f;
+		int32_t raw_temp = (int16_t)(mpu6050_rec_buf[IDX_TEMP_OUT_H] << 8 | mpu6050_rec_buf[IDX_TEMP_OUT_L]);
+		float temperature = 35.0f + ((float)raw_temp + 512.0f) / 340.0f;
 
 		// Apply sensor scaling
 		float gyro_scale = PIOS_MPU6050_GetGyroScale();
@@ -562,7 +562,7 @@ static void PIOS_MPU6050_Task(void* parameters)
 		gyro_data.temperature = temperature;
 
 		portBASE_TYPE xHigherPriorityTaskWoken_gyro;
-		xQueueSendToBackFromISR(dev->gyro_queue, (void*) &gyro_data, &xHigherPriorityTaskWoken_gyro);
+		xQueueSendToBackFromISR(dev->gyro_queue, (void*)&gyro_data, &xHigherPriorityTaskWoken_gyro);
 
 #endif
 	}
